+++ Shiver/Data/Spell.sql
@@ -1,0 +1,23 @@
+Create Table SpellsUsingAmmo (
+	English VarChar(450) Not Null,
+	Constraint PK_SpellsUsingAmmo Primary Key (English),
+);
+/*
+Insert Into SpellsUsingAmmo
+
+-- Turtle 1.17
+Trueshot
+
+-- Vanilla 1.12
+Aimed Shot
+Arcane Shot
+Auto Shot
+Concussive Shot
+Multi-Shot
+Scatter Shot
+Scorpid Sting
+Serpent Sting
+Tranquilizing Shot
+Viper Sting
+Wyvern Sting
+*/